refactor GlassBaseActivity
