+++ collaboration-suite-glass-ui/src/main/java/com/kaleyra/collaboration_suite_glass_ui/chat/GlassChatActivity.kt
@@ -32,20 +32,16 @@ import kotlinx.coroutines.flow.launchIn
 import kotlinx.coroutines.flow.onEach
 import kotlinx.coroutines.launch
 
-internal class GlassChatActivity : GlassBaseActivity(), OnDestinationChangedListener,
-                                   GlassTouchEventManager.Listener {
+internal class GlassChatActivity : GlassBaseActivity() {
 
     private lateinit var binding: KaleyraChatActivityGlassBinding
 
     private val viewModel: ChatViewModel by viewModels()
 
-    private var glassTouchEventManager: GlassTouchEventManager? = null
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = KaleyraChatActivityGlassBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        glassTouchEventManager = GlassTouchEventManager(this, this)
         if (DeviceUtils.isSmartGlass) enableImmersiveMode()
         turnScreenOn()
 
@@ -96,32 +92,10 @@ internal class GlassChatActivity : GlassBaseActivity(), OnDestinationChangedList
     override fun onDestroy() {
         super.onDestroy()
         turnScreenOff()
-        glassTouchEventManager = null
         sendCustomNotificationBroadcast(DisplayedChatActivity.ACTION_CHAT_CLOSE)
     }
 
     override fun onDestinationChanged(destinationId: Int) = Unit
-
-    /**
-     * @suppress
-     */
-    override fun dispatchTouchEvent(ev: MotionEvent?): Boolean =
-        if (glassTouchEventManager!!.toGlassTouchEvent(ev)) true
-        else super.dispatchTouchEvent(ev)
-
-    /**
-     * @suppress
-     */
-    override fun dispatchKeyEvent(event: KeyEvent?): Boolean =
-        if (glassTouchEventManager!!.toGlassTouchEvent(event)) true
-        else super.dispatchKeyEvent(event)
-
-    override fun onGlassTouchEvent(glassEvent: TouchEvent): Boolean {
-        val currentDest =
-            supportFragmentManager.currentNavigationFragment as? TouchEventListener
-                ?: return false
-        return currentDest.onTouch(glassEvent)
-    }
 
     private fun onNewChatIntent(intent: Intent) {
         val userId = intent.extras?.getString("userId") ?: return
